+++ src/test/java/org/folio/entlinks/service/links/LinksSuggestionsServiceTest.java
@@ -24,6 +24,7 @@ import org.folio.entlinks.integration.dto.FieldParsedContent;
 import org.folio.entlinks.integration.dto.SourceParsedContent;
 import org.folio.entlinks.integration.internal.AuthoritySourceFilesService;
 import org.folio.spring.test.type.UnitTest;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -59,7 +60,7 @@ class LinksSuggestionsServiceTest {
     when(sourceFilesService.fetchAuthoritySources()).thenReturn(Map.of(sourceFile.id(), sourceFile));
 
     linksSuggestionService
-      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield);
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, false);
 
     var bibField = bib.getFields().get(0);
     var linkDetails = bibField.getLinkDetails();
@@ -78,6 +79,36 @@ class LinksSuggestionsServiceTest {
 
   @ParameterizedTest
   @ValueSource(strings = {NATURAL_ID_SUBFIELD, ID_SUBFIELD})
+  void fillLinkDetailsWithSuggestedAuthorities_shouldUpdateAndRemoveControlledSubfield(String linkingMatchSubfield) {
+    var rules = getMapRule("100", "100");
+    var initialBibSubfields = new HashMap<String, List<String>>();
+    initialBibSubfields.put("c", List.of("c value"));
+    var bib = getBibParsedRecordContent("100", initialBibSubfields, null);
+    var authority = getAuthorityParsedRecordContent("100");
+    var sourceFile = new AuthoritySourceFile(SOURCE_FILE_ID, BASE_URL, SOURCE_FILE_NAME, codes("e1"));
+
+    when(sourceFilesService.fetchAuthoritySources()).thenReturn(Map.of(sourceFile.id(), sourceFile));
+
+    linksSuggestionService
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, false);
+
+    var bibField = bib.getFields().get(0);
+    var linkDetails = bibField.getLinkDetails();
+    assertEquals(LinkStatus.NEW, linkDetails.getStatus());
+    assertEquals(AUTHORITY_ID, linkDetails.getAuthorityId());
+    assertEquals(NATURAL_ID, linkDetails.getAuthorityNaturalId());
+    assertEquals(1, linkDetails.getLinkingRuleId());
+    assertNull(linkDetails.getErrorCause());
+
+    var bibSubfields = bibField.getSubfields();
+    assertEquals(AUTHORITY_ID.toString(), bibSubfields.get("9").get(0));
+    assertEquals(BASE_URL + NATURAL_ID, bibSubfields.get("0").get(0));
+    assertFalse(bibSubfields.containsKey("c"));
+    assertTrue(bibSubfields.containsKey("b"));
+  }
+
+  @ParameterizedTest
+  @ValueSource(strings = {NATURAL_ID_SUBFIELD, ID_SUBFIELD})
   void fillLinkDetailsWithSuggestedAuthorities_shouldFillLinkDetails_withActualLink(String linkingMatchSubfield) {
     var rules = getMapRule("100", "100");
     var bib = getBibParsedRecordContent("100", getActualLinksDetails());
@@ -87,7 +118,7 @@ class LinksSuggestionsServiceTest {
     when(sourceFilesService.fetchAuthoritySources()).thenReturn(Map.of(sourceFile.id(), sourceFile));
 
     linksSuggestionService
-      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield);
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, false);
 
     var bibField = bib.getFields().get(0);
     var linkDetails = bibField.getLinkDetails();
@@ -113,7 +144,7 @@ class LinksSuggestionsServiceTest {
     var authority = getAuthorityParsedRecordContent("100", emptyMap());
 
     linksSuggestionService
-      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield);
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, false);
 
     var linkDetails = bib.getFields().get(0).getLinkDetails();
     assertEquals(LinkStatus.ERROR, linkDetails.getStatus());
@@ -131,7 +162,7 @@ class LinksSuggestionsServiceTest {
 
     linksSuggestionService
       .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority, secondAuthority), rules,
-        linkingMatchSubfield);
+        linkingMatchSubfield, false);
 
     var linkDetails = bib.getFields().get(0).getLinkDetails();
     assertEquals(LinkStatus.ERROR, linkDetails.getStatus());
@@ -150,7 +181,7 @@ class LinksSuggestionsServiceTest {
     var authority = getAuthorityParsedRecordContent("110");
 
     linksSuggestionService
-      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield);
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, false);
 
     var linkDetails = bib.getFields().get(0).getLinkDetails();
     assertEquals(LinkStatus.ERROR, linkDetails.getStatus());
@@ -168,11 +199,71 @@ class LinksSuggestionsServiceTest {
     var authority = getAuthorityParsedRecordContent("110");
 
     linksSuggestionService
-      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield);
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, false);
 
     var linkDetails = bib.getFields().get(0).getLinkDetails();
     assertEquals(LinkStatus.ERROR, linkDetails.getStatus());
     assertEquals(DISABLED_AUTO_LINKING.getCode(), linkDetails.getErrorCause());
+  }
+
+  @ParameterizedTest
+  @ValueSource(strings = {NATURAL_ID_SUBFIELD, ID_SUBFIELD})
+  void fillLinkDetailsWithSuggestedAuthorities_shouldFillLinkDetails_onIgnoredAutoLinkingFeature(
+    String linkingMatchSubfield) {
+    var rules = getMapRule("100", "100");
+    disableAutoLinkingFeature(rules.get("100"));
+    var bib = getBibParsedRecordContent("100", getActualLinksDetails());
+    var authority = getAuthorityParsedRecordContent("100");
+    var sourceFile = new AuthoritySourceFile(SOURCE_FILE_ID, BASE_URL, SOURCE_FILE_NAME, codes("e1"));
+
+    when(sourceFilesService.fetchAuthoritySources()).thenReturn(Map.of(sourceFile.id(), sourceFile));
+
+    linksSuggestionService
+      .fillLinkDetailsWithSuggestedAuthorities(List.of(bib), List.of(authority), rules, linkingMatchSubfield, true);
+
+    var bibField = bib.getFields().get(0);
+    var linkDetails = bibField.getLinkDetails();
+    assertEquals(LinkStatus.ACTUAL, linkDetails.getStatus());
+    assertEquals(AUTHORITY_ID, linkDetails.getAuthorityId());
+    assertEquals(NATURAL_ID, linkDetails.getAuthorityNaturalId());
+    assertEquals(1, linkDetails.getLinkingRuleId());
+    assertNull(linkDetails.getErrorCause());
+
+    var bibSubfields = bibField.getSubfields();
+    assertEquals(AUTHORITY_ID.toString(), bibSubfields.get("9").get(0));
+    assertEquals(BASE_URL + NATURAL_ID, bibSubfields.get("0").get(0));
+    assertFalse(bibSubfields.containsKey("a"));
+    assertTrue(bibSubfields.containsKey("b"));
+  }
+
+  @Test
+  void shouldFillErrorDetailsWithNoSuggestions_onlyForFieldWithNoError() {
+    var bibs = List.of(getBibParsedRecordContent("100", getActualLinksDetails()),
+        getBibParsedRecordContent("101", getActualLinksDetails().errorCause("test")));
+
+    linksSuggestionService.fillErrorDetailsWithNoSuggestions(bibs, "0");
+
+    assertEquals("101", bibs.get(0).getFields().get(0).getLinkDetails().getErrorCause());
+    assertEquals("test", bibs.get(1).getFields().get(0).getLinkDetails().getErrorCause());
+  }
+
+  @Test
+  void shouldFillErrorDetailsWithDisabledAutoLinking() {
+    var field = new FieldParsedContent("100", "//", "//",
+        Map.of(NATURAL_ID_SUBFIELD, List.of(NATURAL_ID)), null);
+
+    linksSuggestionService.fillErrorDetailsWithDisabledAutoLinking(field, "0");
+
+    assertEquals("103", field.getLinkDetails().getErrorCause());
+  }
+
+  @Test
+  void shouldNotFillErrorDetailsWithDisabledAutoLinking_whenNoSubfield() {
+    var field = new FieldParsedContent("100", "//", "//", Map.of(), null);
+
+    linksSuggestionService.fillErrorDetailsWithDisabledAutoLinking(field, "0");
+
+    assertNull(field.getLinkDetails());
   }
 
   private AuthorityParsedContent getAuthorityParsedRecordContent(String authorityField) {
@@ -186,7 +277,11 @@ class LinksSuggestionsServiceTest {
   }
 
   private SourceParsedContent getBibParsedRecordContent(String bibField, LinkDetails linkDetails) {
-    var subfields = new HashMap<String, List<String>>();
+    return getBibParsedRecordContent(bibField, new HashMap<>(), linkDetails);
+  }
+
+  private SourceParsedContent getBibParsedRecordContent(String bibField, Map<String, List<String>> subfields,
+                                                        LinkDetails linkDetails) {
     subfields.put(NATURAL_ID_SUBFIELD, List.of(NATURAL_ID));
     subfields.put(ID_SUBFIELD, List.of(AUTHORITY_ID.toString()));
     var field = new FieldParsedContent(bibField, "//", "//", subfields, linkDetails);
@@ -209,7 +304,7 @@ class LinksSuggestionsServiceTest {
     rule.setBibField(bibField);
     rule.setAuthorityField(authorityField);
     rule.setAutoLinkingEnabled(true);
-    rule.setAuthoritySubfields(new char[] {'a', 'b'});
+    rule.setAuthoritySubfields(new char[] {'a', 'c'});
     rule.setSubfieldModifications(List.of(modification));
     rule.setSubfieldsExistenceValidations(existence);